refactor: 프론트 디렉토리 정리
