+++ frontend/sin/src/components/api/event/FindEventsApi.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import axios from "axios";
-import {BACKEND_ADDRESS} from "../../constants/ADDRESS";
+import {BACKEND_ADDRESS} from "../../../constants/ADDRESS";
 
 const findEventsApi = () => {
   return (axios.get(BACKEND_ADDRESS + "/goods/event")
