+++ frontend/sin/src/components/api/notice/FindFaqApi.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import axios from "axios";
-import {BACKEND_ADDRESS} from "../../constants/ADDRESS";
+import {BACKEND_ADDRESS} from "../../../constants/ADDRESS";
 const findFaqApi = (page) => {
 
   return page?(axios.get(BACKEND_ADDRESS + "/service/faq?page="+page)
