+++ frontend/sin/src/components/api/notice/GetNoticeApi.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import axios from 'axios';
-import { BACKEND_ADDRESS } from '../../constants/ADDRESS';
+import { BACKEND_ADDRESS } from '../../../constants/ADDRESS';
 
 const getNoticeApi = (page) => {
     return (axios.get(BACKEND_ADDRESS+`/board/list?id=notice&page=`+page)
