+++ frontend/sin/src/components/api/notice/GetNoticeDetailsApi.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import axios from 'axios';
-import { BACKEND_ADDRESS } from '../../constants/ADDRESS';
+import { BACKEND_ADDRESS } from '../../../constants/ADDRESS';
 
 const getNoticeDetailsApi = (no) => {
   return (axios.get(BACKEND_ADDRESS+"/board/view?id=notice&no="+no)
