+++ frontend/sin/src/components/api/notice/GetSearchedNoticeApi.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import axios from 'axios';
-import { BACKEND_ADDRESS } from '../../constants/ADDRESS';
+import { BACKEND_ADDRESS } from '../../../constants/ADDRESS';
 
 const GetSearchedNoticeApi = (url,word,page) => {
     return (
