+++ frontend/sin/src/components/api/product/FindProductListApi.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import axios from "axios";
-import {BACKEND_ADDRESS} from "../../constants/ADDRESS";
+import {BACKEND_ADDRESS} from "../../../constants/ADDRESS";
 
 const findProductListApi = (category, list) => {
   return category ? axios.get(BACKEND_ADDRESS + "/goods/goods_list?category=" + category).then(response => response.data)
