+++ frontend/sin/src/components/notice/Noticesection.js
@@ -3,8 +3,8 @@ import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 import axios from 'axios';
 import qs from 'query-string';
-import getNoticeApi from '../api/GetNoticeApi';
-import GetSearchedNoticeApi from '../api/GetSearchedNoticeApi';
+import getNoticeApi from '../api/notice/GetNoticeApi';
+import GetSearchedNoticeApi from '../api/notice/GetSearchedNoticeApi';
 
 const Mainrighttitle = styled.div`padding: 20px 0;`
 const Mainrighttitlestrong = styled.strong`display: inline-block; font-size: 24px; font-weight: bold; padding-right: 10px;`
